requirements.txt
anyio==4.0.0
arabic-reshaper==3.0.0
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.13.3
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==44.0.0
Django==4.2.5
django-cors-headers==4.2.0
django-crispy-forms==2.4
django-dbbackup==4.1.0
django-filter==23.3
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-history==1.0.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==0.18.0
httpx==0.25.0
idna==3.4
numpy==2.0.2
openpyxl==3.1.2
packaging==26.0
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.8
pycparser==2.22
PyJWT==2.8.0
PyPDF2==3.0.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.7
reportlab==3.6.6
requests==2.32.5
dj-database-url==2.2.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.6
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.5
